backend/requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Deprecated==1.2.13
Django==3.2.12
django-cors-headers==3.11.0
django-environ==0.8.1
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
google-auth==2.6.2
idna==3.3
jwcrypto==1.0
oauthlib==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
requests==2.27.1
rsa==4.8
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
